requirements.txt
# Property Voice Agent - Dependencies
# FastAPI & Server
fastapi==0.110.0
uvicorn[standard]==0.27.0
python-multipart==0.0.9

# Database
supabase==2.3.0

# Email
resend==0.8.0

# AI
openai==1.12.0

# Utilities
python-dotenv==1.0.0
pydantic==2.6.0
pydantic-settings==2.1.0
requests==2.31.0
httpx>=0.24.0,<0.25.0

# Development
pytest==8.0.0
pytest-asyncio==0.23.0
